Add files via upload
Added for loop that has a variable "target_theta" depending on where the sensor is seeing the line. This is done with the "(3-j)*3". If adc1[6] (pin 12) is reading the line, then 3-6 = -3 so it will do a hard turn of -3 degrees to the right. If adc1[0] (pin 0 I think) is reading the line, then 3-0 = 3 so it will do a hard turn 3 degrees to the left to keep with the line. I have not run this on the mouse yet, so there may need to be some trial and error to see if this can work.